split rendering into its own file
